// src/app/Models/Pagos/ProcesarMovimientoPagoPinPadMovilRequest.ts
export class  ProcesarMovimientoPagoPinPadMovilRequest {
  folioOperacionAsociada: string;
  codigoFormaPagoImporte: string;
  importeVentaTotal: number;
  estatus: string;
  secuenciaFormaPagoImporte: number;
  folioAutorizacionPinPadMovil: string;
  descuentosPromocionalesPorVentaAplicados: any;
  descuentosPromocionalesPorLineaAplicados: any;
}